Change project structure of frontend, move lib out of app dir
