+++ frontend/src/lib/components/Logo.tsx
