+++ frontend/src/lib/services/NetworkService.ts
